% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bivariate_density_c}
\alias{bivariate_density_c}
\title{A Gibbs sampler using Rcpp}
\usage{
bivariate_density_c(a, b, n)
}
\arguments{
\item{a}{fixed number for beta distribution}

\item{b}{fixed number for beta distribution}

\item{n}{fixed number for binomial distribution}
}
\value{
a binary Monte Carlo Markov Chain
}
\description{
A Gibbs sampler using Rcpp to generate a chain with target joint density
}
