% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21075.R
\name{coef_lasso}
\alias{coef_lasso}
\title{LASSO estimates}
\usage{
coef_lasso(y, x)
}
\arguments{
\item{y}{vector of response}

\item{x}{design matrix}
}
\value{
the lasso estimator of coefficients
}
\description{
This code estimates coefficient for the LASSO procedure.
}
