% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21075.R
\name{ME}
\alias{ME}
\title{model error (ME)}
\usage{
ME(beta, betahat, x)
}
\arguments{
\item{beta}{true coefficient}

\item{betahat}{the estimated coeffcient from a specific method}

\item{x}{the design matrix}
}
\value{
the model error
}
\description{
model error (ME) for prediction accuracy
}
