\name{StatComp21075-package}
\alias{StatComp21075-package}
\docType{package}
\title{
  A PACS procedure for variable selection.
}
\description{
  A PACS procedure for variable selection,comparision with LASSO estimates using medel error function.
}
\author{
Yu Zheng.

Maintainer: Yu Zheng <belzheng@mail.ustc.edu.cn>
}
\keyword{ package }
